Making it not care about the contents of a stack_yaml
Parsing the cloudformation yaml is less trivial than parsing json and
Bespin doesn't actually care for the contents of a stack.

So, we make it read in the stack as just a string for stack_yaml and
pass around that string.

We still treat stack_json as a json object so that we know it's valid
json....