Do not merge extra_files over main bespin.yml
Currently the collector adds bespin.yml configuration and then merges
extra_files over the top (in order). This functionality is counter
intuitive to new users as the primary configuration should be a source
of truth. YAML entries defined in the primary configuration should
override all others.

Change the load of additional configuration files (if specified) to
happen prior to the main bespin.yml